@@ -1,8 +1,6 @@
 package com.example.inventoryservice.service;
 
-import com.example.inventoryservice.dto.InventoryDto;
-import com.example.inventoryservice.model.Inventory;
-import com.example.inventoryservice.repository.request.InventoryRequest;
+import com.example.inventoryservice.request.InventoryRequest;
 import com.example.inventoryservice.response.ApiResponse;
 
 public interface InventoryService {
