@@ -13,7 +13,7 @@ import org.springframework.web.bind.annotation.*;
 
 
 @RestController
-@RequestMapping("/com/api/products")
+@RequestMapping("/com/api/product-service")
 public class ProductController {
 
     private static final Logger logger = LoggerFactory.getLogger(ProductController.class);
@@ -21,25 +21,32 @@ public class ProductController {
     @Autowired
     private ProductService productService;
 
-    @PostMapping
+    @PostMapping("/products")
     public ApiResponse createProduct(@Valid @RequestBody ProductRequest product) {
         logger.info("Request to create product: {}", product.getName());
         return productService.createProduct(product);
     }
 
-    @PutMapping("/{id}")
-    public ApiResponse updateProduct(@PathVariable Long id, @RequestBody ProductRequest product) {
+    @PutMapping("/product")
+    public ApiResponse updateProduct(@RequestParam Long id, @RequestBody ProductRequest product) {
         logger.info("Request to update product with ID: {}", id);
         return productService.updateProduct(id, product);
     }
 
-    @GetMapping("/getByProductId")
-    public ApiResponse getProductById(@RequestParam Long id) {
-        logger.info("Request to retrieve product with ID: {}", id);
-        return productService.getProductById(id);
+    @GetMapping("/product")
+    public ApiResponse getProductById(@RequestParam("id") Long productId,
+                                      @RequestHeader(value = "Service-Name", required = false) String serviceName) {
+
+        if (serviceName != null) {
+            logger.info("Request received from Service: {}", serviceName);
+        } else {
+            logger.info("Request received from an external client.");
+        }
+
+        return productService.getProductById(productId);
     }
 
-    @GetMapping
+    @GetMapping("/products")
     public ApiResponse getAllProducts() {
         logger.info("Request to retrieve all products");
         return productService.getAllProducts();
