@@ -1,44 +1,18 @@
 package com.example.productservice.dto;
 
-import com.example.productservice.model.Category;
-import com.fasterxml.jackson.annotation.JsonBackReference;
-import com.fasterxml.jackson.annotation.JsonInclude;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.validation.constraints.*;
-import lombok.*;
-
-import java.time.LocalDateTime;
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
 
 @Data
-@AllArgsConstructor
 @NoArgsConstructor
-@JsonInclude(JsonInclude.Include.NON_NULL)
+@AllArgsConstructor
 public class ProductDto {
     private Long id;
-
-    @NotBlank(message = "Product name is required")
-    @Size(max = 100, message = "Product name cannot exceed 100 characters")
     private String name;
-
-    @NotBlank(message = "Brand is required")
-    @Size(max = 50, message = "Brand name cannot exceed 50 characters")
     private String brand;
-
-    @NotBlank(message = "Description is required")
-    @Size(max = 500, message = "Description cannot exceed 500 characters")
     private String description;
-
-    @NotNull(message = "Price is required")
-    @DecimalMin(value = "0.0", inclusive = true, message = "Price cannot be negative")
-    private Double price;
-
-    @NotNull(message = "Quantity is required")
-    @Min(value = 0, message = "Quantity cannot be negative")
-    private Integer quantity;
-
-    private Category category;
-
+    private double price;
+    private String quantityStatus; // "In Stock" or "Out of Stock"
+    private String category; // Add category field
 }
-
