@@ -27,7 +27,7 @@ public class SecurityConfig {
     public SecurityFilterChain securityFilterChain(HttpSecurity http) throws Exception {
         return http.csrf(AbstractHttpConfigurer::disable)
                 .authorizeHttpRequests(auth -> auth
-                        .requestMatchers("/com/api/user","/com/api/login","/swagger-ui/**", "/v3/api-docs/**").permitAll()
+                        .requestMatchers("/com/api/user-service/users","/com/api/user-service/login","/swagger-ui/**", "/v3/api-docs/**","/com/api/user-service/validate").permitAll()
                         .anyRequest().authenticated())
                 .sessionManagement(session -> session.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
                 .addFilterBefore(jwtFilter, UsernamePasswordAuthenticationFilter.class)
