@@ -1,7 +1,6 @@
 package com.example.userservice.service;
 
 import com.example.userservice.model.Role;
-import com.example.userservice.model.User;
 import com.example.userservice.request.UserRequest;
 import com.example.userservice.response.ApiResponse;
 import jakarta.servlet.http.HttpSession;
@@ -10,7 +9,8 @@ public interface UserService {
  ApiResponse createUser (UserRequest user);
  ApiResponse getUserById(Long userId);
  ApiResponse createRole(Role role);
- ApiResponse verify(String email, String password, HttpSession httpSession);
+ ApiResponse login(String email, String password, HttpSession httpSession);
  ApiResponse updateUser(Long userId,UserRequest user);
  ApiResponse deleteUser(Long userId);
+ ApiResponse validateToken( String token);
 }
